UPDATE + Updated Universe background + Planets have their orbit lanes again + Planets rotate and around the sun and their own orbit + Window size is now automatically determined by screen resolution + Window is now automatically positioned in the center of the screen. + Minor code clean up

+++ OpenGL/Axes.h
@@ -12,8 +12,8 @@ class Axes
 private:
 	bool mIsXBeingRendered , mIsYBeingRendered, mIsZBeingRendered;
 	bool mIsXGridBeingRendered, mIsYGridBeingRendered, mIsZGridBeingRendered;
-	int mLengthOfAxes;
-	int depth;
+	GLfloat mLengthOfAxes;
+	GLfloat depth;
 	int squareSizeInGrid;
 
 public:
